rename Components directory to components
