+++ client/src/components/DiscordLog/DiscordLog.scss
