+++ client/src/components/DiscordLog/DiscordLog.tsx
