+++ client/src/components/DiscordLog/DiscordLogEntry.scss
