+++ client/src/components/DiscordLog/DiscordLogEntry.tsx
