+++ client/src/components/Events/EventEntry.scss
