+++ client/src/components/Events/EventPage.scss
