+++ client/src/components/Log/Log.scss
