+++ client/src/components/Log/Log.tsx
