+++ client/src/components/Log/LogEntry.scss
