+++ client/src/components/Log/LogEntry.tsx
