+++ client/src/components/Roster/GuildMemberCard.scss
