+++ client/src/components/Roster/GuildMemberDetails.scss
