+++ client/src/components/Roster/GuildMemberDetails.tsx
