+++ client/src/components/Roster/KickModal.scss
