+++ client/src/components/Roster/Warnings/WarningsViewer.scss
