+++ client/src/lib/apis/recruitment-api.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosError } from 'axios';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { useToast } from '../../Components/Common/ToastContext';
+import { useToast } from '../../components/Common/ToastContext';
 import { RecruitmentPost } from '../interfaces/recruitment-post';
 
 export const useRecruitmentPost = () => useQuery<RecruitmentPost, AxiosError>('recruitment-post');
